refactor(parser): split stmt/control.rs into cohesive submodules (§7-8)
Split the 2941-line parser/stmt/control.rs (over the 500-line limit) into
a facade plus ten themed submodules under src/parser/stmt/control/, all
under 500 lines. control.rs is entirely free parser-combinator functions
addressed as control::<name> (27 call sites); the facade re-exports each
submodule preserving its original visibility, so no external caller
changes. stmt/mod.rs is untouched (submodules declared inside the facade).

- control.rs (208)          - facade: use lines, shared helpers, re-exports
- with_stmt.rs (459)        - with/without/orwith
- loop_repeat.rs (444)      - loop / repeat
- pointy_param.rs (436)     - parse_pointy_param
- for_params.rs (401)       - parse_for_params
- conditionals.rs (398)     - if/unless + elsif-chain lowering
- for_loops.rs (256)        - for/foreach/hyper/race/lazy entry points
- labeled_loop.rs (169)     - labeled loops
- while_until.rs (131)      - while / until
- react.rs (44)             - react / whenever
- given_when.rs (39)        - given/when/default

Pure relocation. Moved public fns are pub(crate) in the submodules and
re-exported at their original visibility (pub(super) / pub(in crate::parser))
through the facade; shared private helpers stay in the facade. The fn
multiset is byte-identical. No behavior change.
