+++ src/parser/stmt/control.rs
@@ -11,21 +11,6 @@ use crate::value::Value;
 use super::decl::parse_array_shape_suffix;
 use super::sub;
 use super::{block, block_inner, ident, keyword, parse_comma_or_expr, statement, var_name};
-
-static IF_BIND_TMP_COUNTER: AtomicUsize = AtomicUsize::new(0);
-
-#[derive(Clone)]
-struct IfChainClause {
-    cond: Expr,
-    then_branch: Vec<Stmt>,
-    binding_var: Option<String>,
-}
-
-struct ElseClause {
-    binding_params: Option<Vec<ParamDef>>,
-    body: Vec<Stmt>,
-}
-
 fn condition_has_assignment_tail(rest: &str) -> bool {
     if rest.starts_with(":=") || rest.starts_with("::=") || rest.starts_with("⚛=") {
         return true;
@@ -83,1001 +68,6 @@ fn condition_expr(input: &str) -> PResult<'_, Expr> {
     }
 }
 
-fn conditional_expr(input: &str) -> PResult<'_, Expr> {
-    match parse_comma_or_expr(input) {
-        Ok((rest, cond)) => {
-            let (tail, _) = ws(rest)?;
-            if condition_has_assignment_tail(tail)
-                && let Ok((assign_rest, assign_cond)) = super::assign::try_parse_assign_expr(input)
-            {
-                return Ok((assign_rest, assign_cond));
-            }
-            Ok((rest, cond))
-        }
-        Err(_) => condition_expr(input),
-    }
-}
-
-pub(super) fn if_stmt(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("if", input).ok_or_else(|| PError::expected("if statement"))?;
-    let (rest, _) = ws1(rest)?;
-    let (rest, cond) = conditional_expr(rest)?;
-    let (rest, _) = ws(rest)?;
-    let (rest, binding_params) = parse_if_binding_params(rest)?;
-    let (rest, _) = ws(rest)?;
-    let (rest, raw_then_branch) = block(rest)?;
-    let (rest, _) = ws(rest)?;
-    let (binding_var, then_branch) = lower_if_clause_binding(binding_params, raw_then_branch);
-
-    let mut clauses = vec![IfChainClause {
-        cond,
-        then_branch,
-        binding_var,
-    }];
-    let (rest, (mut elsif_clauses, else_clause)) = parse_elsif_chain(rest)?;
-    clauses.append(&mut elsif_clauses);
-
-    let stmt = lower_if_chain(clauses, else_clause);
-    Ok((rest, stmt))
-}
-
-fn parse_if_binding_params(input: &str) -> PResult<'_, Option<Vec<ParamDef>>> {
-    let Some(rest) = input.strip_prefix("->") else {
-        return Ok((input, None));
-    };
-    let (rest, _) = ws(rest)?;
-    // Zero-parameter pointy block: `if EXPR -> { ... }`
-    if rest.starts_with('{') {
-        return Ok((rest, Some(Vec::new())));
-    }
-
-    let (rest, params) = if let Some(rest) = rest.strip_prefix('(') {
-        let (rest, _) = ws(rest)?;
-        let (rest, params) = super::parse_param_list_pub(rest)?;
-        let (rest, _) = ws(rest)?;
-        let (rest, _) = parse_char(rest, ')')?;
-        (rest, params)
-    } else {
-        super::parse_param_list_pub(rest)?
-    };
-    let (rest, _) = ws(rest)?;
-    let (rest, _) = if let Some(after_arrow) = rest.strip_prefix("-->") {
-        let (rest, _) = super::parse_return_type_annotation_pub(after_arrow)?;
-        let (rest, _) = ws(rest)?;
-        (rest, ())
-    } else {
-        (rest, ())
-    };
-    Ok((rest, Some(params)))
-}
-
-fn is_simple_if_binding(param: &ParamDef) -> bool {
-    param.traits.is_empty()
-        && param.shape_constraints.is_none()
-        && !param.named
-        && !param.slurpy
-        && !param.double_slurpy
-        && param.default.is_none()
-        && !param.optional_marker
-        && param.type_constraint.is_none()
-        && param.sub_signature.is_none()
-        && param.outer_sub_signature.is_none()
-        && param.code_signature.is_none()
-}
-
-fn next_if_bind_tmp_name() -> String {
-    let tmp_idx = IF_BIND_TMP_COUNTER.fetch_add(1, Ordering::Relaxed);
-    format!("$__mutsu_if_bind_{tmp_idx}")
-}
-
-fn lower_if_clause_binding(
-    binding_params: Option<Vec<ParamDef>>,
-    then_branch: Vec<Stmt>,
-) -> (Option<String>, Vec<Stmt>) {
-    let Some(param_defs) = binding_params else {
-        return (None, then_branch);
-    };
-    if param_defs.is_empty() {
-        return (None, then_branch);
-    }
-    if param_defs.len() == 1 && is_simple_if_binding(&param_defs[0]) {
-        return (Some(param_defs[0].name.clone()), then_branch);
-    }
-
-    let source_binding = next_if_bind_tmp_name();
-    let source_expr = Expr::Var(source_binding.trim_start_matches('$').to_string());
-    // For **@ (double slurpy) as the only positional param, pass the condition
-    // as a single argument without slipping, so the list is captured as-is.
-    let has_only_double_slurpy = param_defs
-        .iter()
-        .filter(|p| !p.named)
-        .all(|p| p.double_slurpy);
-    let args = if has_only_double_slurpy {
-        vec![source_expr]
-    } else {
-        vec![Expr::Unary {
-            op: TokenKind::Pipe,
-            expr: Box::new(source_expr),
-        }]
-    };
-    let call_expr = Expr::CallOn {
-        target: Box::new(Expr::AnonSubParams {
-            params: param_defs.iter().map(|p| p.name.clone()).collect(),
-            param_defs,
-            return_type: None,
-            body: then_branch,
-            is_rw: false,
-            is_whatever_code: false,
-        }),
-        args,
-    };
-    (Some(source_binding), vec![Stmt::Expr(call_expr)])
-}
-
-fn ensure_last_clause_binding_var(clauses: &mut [IfChainClause]) -> Option<String> {
-    let last_clause = clauses.last_mut()?;
-    Some(if let Some(existing) = &last_clause.binding_var {
-        existing.clone()
-    } else {
-        let generated = next_if_bind_tmp_name();
-        last_clause.binding_var = Some(generated.clone());
-        generated
-    })
-}
-
-fn lower_else_binding(source_binding: &str, else_clause: ElseClause) -> Vec<Stmt> {
-    let Some(param_defs) = else_clause.binding_params else {
-        return else_clause.body;
-    };
-    if param_defs.is_empty() {
-        return else_clause.body;
-    }
-    if param_defs.len() == 1 && is_simple_if_binding(&param_defs[0]) {
-        let mut body = Vec::with_capacity(else_clause.body.len() + 1);
-        body.push(Stmt::VarDecl {
-            name: param_defs[0].name.clone(),
-            expr: Expr::Var(source_binding.trim_start_matches('$').to_string()),
-            type_constraint: None,
-            is_state: false,
-            is_our: false,
-            is_dynamic: false,
-            is_export: false,
-            export_tags: Vec::new(),
-            custom_traits: Vec::new(),
-            where_constraint: None,
-        });
-        body.extend(else_clause.body);
-        return body;
-    }
-
-    let call_expr = Expr::CallOn {
-        target: Box::new(Expr::AnonSubParams {
-            params: param_defs.iter().map(|p| p.name.clone()).collect(),
-            param_defs,
-            return_type: None,
-            body: else_clause.body,
-            is_rw: false,
-            is_whatever_code: false,
-        }),
-        args: vec![Expr::Unary {
-            op: TokenKind::Pipe,
-            expr: Box::new(Expr::Var(
-                source_binding.trim_start_matches('$').to_string(),
-            )),
-        }],
-    };
-    vec![Stmt::Expr(call_expr)]
-}
-
-fn parse_elsif_chain(input: &str) -> PResult<'_, (Vec<IfChainClause>, Option<ElseClause>)> {
-    let mut rest = input;
-    let mut clauses = Vec::new();
-    let mut last_orwith_cond: Option<Expr> = None;
-
-    loop {
-        if let Some(r) = keyword("elsif", rest) {
-            let (r, _) = ws1(r)?;
-            let (r, cond) = conditional_expr(r)?;
-            let (r, _) = ws(r)?;
-            let (r, binding_params) = parse_if_binding_params(r)?;
-            let (r, _) = ws(r)?;
-            let (r, raw_then_branch) = block(r)?;
-            let (r, _) = ws(r)?;
-            let (binding_var, then_branch) =
-                lower_if_clause_binding(binding_params, raw_then_branch);
-            clauses.push(IfChainClause {
-                cond,
-                then_branch,
-                binding_var,
-            });
-            last_orwith_cond = None;
-            rest = r;
-            continue;
-        }
-        if let Some(r) = keyword("orwith", rest) {
-            let (r, _) = ws1(r)?;
-            let (r, orwith_cond_expr) = condition_expr(r)?;
-            let (r, _) = ws(r)?;
-            // Check for optional pointy block: orwith EXPR -> $param { ... }
-            let (r, orwith_param_name) = if let Some(r2) = r.strip_prefix("->") {
-                let (r2, _) = ws(r2)?;
-                if let Some(r_after_sigil) = r2.strip_prefix('$') {
-                    let end = r_after_sigil
-                        .find(|c: char| !c.is_alphanumeric() && c != '_' && c != '-')
-                        .unwrap_or(r_after_sigil.len());
-                    let name = &r_after_sigil[..end];
-                    let r2 = &r_after_sigil[end..];
-                    let (r2, _) = ws(r2)?;
-                    (r2, Some(name.to_string()))
-                } else {
-                    (r, None)
-                }
-            } else {
-                (r, None)
-            };
-            let (r, orwith_body) = block(r)?;
-            // Topicalize $_ and optional param in the orwith body
-            let mut orwith_then = vec![topicalize(&orwith_cond_expr)];
-            if let Some(ref pname) = orwith_param_name {
-                orwith_then.push(Stmt::VarDecl {
-                    name: pname.clone(),
-                    expr: orwith_cond_expr.clone(),
-                    type_constraint: None,
-                    is_state: false,
-                    is_our: false,
-                    is_dynamic: false,
-                    is_export: false,
-                    export_tags: Vec::new(),
-                    custom_traits: Vec::new(),
-                    where_constraint: None,
-                });
-            }
-            orwith_then.extend(orwith_body);
-            // orwith uses .defined as the condition
-            last_orwith_cond = Some(orwith_cond_expr.clone());
-            let orwith_cond = Expr::MethodCall {
-                target: Box::new(orwith_cond_expr),
-                name: Symbol::intern("defined"),
-                args: Vec::new(),
-                modifier: None,
-                quoted: false,
-            };
-            let (r, _) = ws(r)?;
-            clauses.push(IfChainClause {
-                cond: orwith_cond,
-                then_branch: orwith_then,
-                binding_var: None,
-            });
-            rest = r;
-            continue;
-        }
-        break;
-    }
-
-    if let Some(r) = keyword("else", rest) {
-        let (r, _) = ws(r)?;
-        let (r, binding_params) = parse_if_binding_params(r)?;
-        let (r, _) = ws(r)?;
-        let (r, mut body) = block(r)?;
-        // If the last clause was `orwith`, topicalize $_ in the else body
-        if let Some(ref orwith_expr) = last_orwith_cond {
-            let mut topicalized = vec![topicalize(orwith_expr)];
-            topicalized.append(&mut body);
-            body = topicalized;
-        }
-        return Ok((
-            r,
-            (
-                clauses,
-                Some(ElseClause {
-                    binding_params,
-                    body,
-                }),
-            ),
-        ));
-    }
-
-    Ok((rest, (clauses, None)))
-}
-
-fn lower_if_chain(mut clauses: Vec<IfChainClause>, else_clause: Option<ElseClause>) -> Stmt {
-    let mut else_branch = if let Some(else_clause) = else_clause {
-        let mut body = lower_else_clause(&mut clauses, else_clause);
-        // An explicit `else {}` with an empty body should evaluate to Nil,
-        // not be indistinguishable from a missing else clause.
-        if body.is_empty() {
-            body.push(Stmt::Expr(Expr::Literal(crate::value::Value::Nil)));
-        }
-        body
-    } else {
-        Vec::new()
-    };
-
-    while let Some(clause) = clauses.pop() {
-        else_branch = vec![Stmt::If {
-            cond: clause.cond,
-            then_branch: clause.then_branch,
-            else_branch,
-            binding_var: clause.binding_var,
-        }];
-    }
-
-    else_branch
-        .pop()
-        .expect("if chain must have at least one clause")
-}
-
-fn lower_else_clause(clauses: &mut [IfChainClause], else_clause: ElseClause) -> Vec<Stmt> {
-    if else_clause.binding_params.is_none() {
-        return else_clause.body;
-    }
-    let Some(source_binding) = ensure_last_clause_binding_var(clauses) else {
-        return else_clause.body;
-    };
-    lower_else_binding(&source_binding, else_clause)
-}
-
-/// Parse `unless` statement.
-pub(super) fn unless_stmt(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("unless", input).ok_or_else(|| PError::expected("unless statement"))?;
-    let (rest, _) = ws1(rest)?;
-    let (rest, cond) = condition_expr(rest)?;
-    let (rest, _) = ws(rest)?;
-    let (rest, body) = block(rest)?;
-    // unless cannot have else/elsif/orwith — check but consume the trailing clause
-    let (check, _) = ws(rest)?;
-    for kw in &["else", "elsif", "orwith"] {
-        if let Some(r) = keyword(kw, check) {
-            // Consume the rest of the invalid clause to produce a hard error
-            let (r, _) = ws(r)?;
-            // Skip condition (if any) and block
-            let r = if kw != &"else" {
-                if let Ok((r, _)) = expression(r) { r } else { r }
-            } else {
-                r
-            };
-            let (r, _) = ws(r)?;
-            let r = if let Ok((r, _)) = block(r) { r } else { r };
-            return Ok((
-                r,
-                Stmt::Die(Expr::Literal(crate::value::Value::str(format!(
-                    "X::Syntax::UnlessElse: unless does not allow '{kw}'"
-                )))),
-            ));
-        }
-    }
-    Ok((
-        rest,
-        Stmt::If {
-            cond: Expr::Unary {
-                op: TokenKind::Bang,
-                expr: Box::new(cond),
-            },
-            then_branch: body,
-            else_branch: Vec::new(),
-            binding_var: None,
-        },
-    ))
-}
-
-/// Parse labeled statement: `LABEL: <statement>`.
-/// Loop-like statements keep the label on the loop node for last/next/redo.
-/// Other statements are wrapped in `Stmt::Label`.
-pub(super) fn labeled_loop_stmt(input: &str) -> PResult<'_, Stmt> {
-    let (rest, label) = ident(input)?;
-    // A statement label requires the colon immediately after the identifier
-    // (`LABEL:`). Whitespace before the colon means this is not a label but a
-    // call with a colonpair/object-hash argument (e.g. `is-deeply :{...}, ...`).
-    if !rest.starts_with(':') || rest.starts_with("::") {
-        return Err(PError::expected("labeled loop"));
-    }
-    let rest = &rest[1..]; // consume ':'
-    let has_space_after_colon = rest.chars().next().is_some_and(char::is_whitespace);
-    if !has_space_after_colon
-        && !rest.starts_with('{')
-        && !rest.starts_with("for")
-        && !rest.starts_with("while")
-        && !rest.starts_with("until")
-        && !rest.starts_with("loop")
-        && !rest.starts_with("do")
-    {
-        return Err(PError::expected("labeled loop"));
-    }
-    let (rest, _) = ws(rest)?;
-
-    // Check which loop keyword follows
-    if let Some(r) = keyword("for", rest) {
-        let (r, _) = ws1(r)?;
-        // Perl 5 `for my $x (LIST) { }` foreach syntax.
-        if looks_like_p5_foreach(r) {
-            return Err(p5_foreach_error());
-        }
-        let (r, iterable) = parse_comma_or_expr(r)?;
-        let (r, _) = ws(r)?;
-        let (r, (param, param_def, params, params_def, rw_block, explicit_zero_params)) =
-            parse_for_params(r)?;
-        let (r, _) = ws(r)?;
-        let (r, body) = block(r)?;
-        return Ok((
-            r,
-            Stmt::For {
-                iterable,
-                param,
-                param_def: Box::new(param_def),
-                params,
-                params_def,
-                body,
-                label: Some(label),
-                mode: crate::ast::ForMode::Normal,
-                rw_block,
-                explicit_zero_params,
-            },
-        ));
-    }
-    if let Some(r) = keyword("while", rest) {
-        let (r, _) = ws1(r)?;
-        let (r, cond) = condition_expr(r)?;
-        let (r, _) = ws(r)?;
-        let (r, body) = block(r)?;
-        return Ok((
-            r,
-            Stmt::While {
-                cond,
-                body,
-                label: Some(label),
-            },
-        ));
-    }
-    if let Some(r) = keyword("until", rest) {
-        let (r, _) = ws1(r)?;
-        let (r, cond) = condition_expr(r)?;
-        let (r, _) = ws(r)?;
-        let (r, body) = block(r)?;
-        return Ok((
-            r,
-            Stmt::While {
-                cond: Expr::Unary {
-                    op: TokenKind::Bang,
-                    expr: Box::new(cond),
-                },
-                body,
-                label: Some(label),
-            },
-        ));
-    }
-    if let Some(r) = keyword("loop", rest) {
-        let (r, _) = ws(r)?;
-        let (r, body) = block(r)?;
-        return Ok((
-            r,
-            Stmt::Loop {
-                init: None,
-                cond: None,
-                step: None,
-                body,
-                repeat: false,
-                label: Some(label),
-            },
-        ));
-    }
-
-    // Label before `do` block: `A: do { ... }`
-    if keyword("do", rest).is_some() {
-        // Parse the rest as a statement and wrap with label
-        // TODO: Represent labeled `do { ... }` with a dedicated AST node instead of
-        // lowering it to a dummy `Stmt::For` carrying `Nil`.
-        let r = keyword("do", rest).unwrap();
-        let (r, _) = ws(r)?;
-        let (r, body) = block(r)?;
-        return Ok((
-            r,
-            Stmt::For {
-                iterable: Expr::ArrayLiteral(vec![Expr::Literal(crate::value::Value::Nil)]),
-                param: None,
-                param_def: Box::new(None),
-                params: Vec::new(),
-                params_def: Vec::new(),
-                body,
-                label: Some(label),
-                mode: crate::ast::ForMode::Normal,
-                rw_block: false,
-                explicit_zero_params: false,
-            },
-        ));
-    }
-    // Label before bare block: `A: { ... }`
-    // TODO: Represent labeled bare blocks directly instead of lowering to
-    // a dummy `Stmt::For` carrying `Nil`.
-    if rest.starts_with('{') {
-        let (r, body) = block(rest)?;
-        return Ok((
-            r,
-            Stmt::For {
-                iterable: Expr::ArrayLiteral(vec![Expr::Literal(crate::value::Value::Nil)]),
-                param: None,
-                param_def: Box::new(None),
-                params: Vec::new(),
-                params_def: Vec::new(),
-                body,
-                label: Some(label),
-                mode: crate::ast::ForMode::Normal,
-                rw_block: false,
-                explicit_zero_params: false,
-            },
-        ));
-    }
-
-    // `foo: args` where `foo` is a declared subroutine is a colon-listop call,
-    // not a label (Raku: `trim: "x"` calls `trim("x")`, while `LOOP: for ...`
-    // is a label). The loop/block label forms are handled above; only the
-    // generic `IDENT: <expr>` case is ambiguous, and a declared-sub name
-    // resolves it to a call. Parse the colon-listop argument list directly.
-    if crate::parser::stmt::simple::is_user_declared_sub(&label) {
-        let (r, call) = crate::parser::primary::ident::parse_expr_listop_args(rest, label.clone())?;
-        return Ok((r, Stmt::Expr(call)));
-    }
-
-    // Generic labeled statement: LABEL: <statement>
-    let (r, stmt) = statement(rest)?;
-    Ok((
-        r,
-        Stmt::Label {
-            name: label,
-            stmt: Box::new(stmt),
-        },
-    ))
-}
-
-/// Parsed for-loop parameter info: `(param, param_def, params, rw_block, explicit_zero_params)`.
-type ForParams = (
-    Option<String>,
-    Option<ParamDef>,
-    Vec<String>,
-    Vec<ParamDef>,
-    bool,
-    bool,
-);
-
-/// Parse for loop parameters: -> $param or -> $a, $b
-/// Returns `(param, param_def, params, rw_block)`.
-/// `rw_block` is `true` when `<->` is used instead of `->`.
-pub(super) fn parse_for_params(input: &str) -> PResult<'_, ForParams> {
-    fn skip_pointy_return_type<'a>(mut r: &'a str) -> PResult<'a, ()> {
-        let (r2, _) = ws(r)?;
-        r = r2;
-        if let Some(after_arrow) = r.strip_prefix("-->") {
-            let (after_arrow, _) = super::parse_return_type_annotation_pub(after_arrow)?;
-            let (after_arrow, _) = ws(after_arrow)?;
-            Ok((after_arrow, ()))
-        } else {
-            Ok((r, ()))
-        }
-    }
-
-    // Check for `<->` (rw pointy block) — must check before `->`
-    let rw_block = input.starts_with("<->");
-    let pointy_stripped = if rw_block {
-        Some(&input[3..])
-    } else {
-        input.strip_prefix("->")
-    };
-    if let Some(stripped) = pointy_stripped {
-        let (r, _) = ws(stripped)?;
-        // Zero-parameter pointy block: for @a -> { ... }
-        // Explicitly declares zero params — passing any arg should throw.
-        if r.starts_with('{') {
-            return Ok((r, (None, None, Vec::new(), Vec::new(), rw_block, true)));
-        }
-        // Parenthesized destructuring pointy param:
-        //   -> ($a, $b) { ... }
-        //   -> (:key($k), :value($v)) { ... }
-        if r.starts_with('(') {
-            let (r, _) = parse_char(r, '(')?;
-            let (r, _) = ws(r)?;
-            let (r, sub_params) = super::parse_param_list_pub(r)?;
-            let (r, _) = ws(r)?;
-            let (r, _) = parse_char(r, ')')?;
-            let (r, _) = skip_pointy_return_type(r)?;
-            let unpack_name = "__for_unpack".to_string();
-            let unpack_def = ParamDef {
-                name: unpack_name.clone(),
-                default: None,
-                multi_invocant: true,
-                required: false,
-                named: false,
-                slurpy: false,
-                sigilless: false,
-                type_constraint: None,
-                literal_value: None,
-                sub_signature: Some(sub_params),
-                where_constraint: None,
-                traits: Vec::new(),
-                double_slurpy: false,
-                onearg: false,
-                optional_marker: false,
-                outer_sub_signature: None,
-                code_signature: None,
-                is_invocant: false,
-                shape_constraints: None,
-            };
-            return Ok((
-                r,
-                (
-                    Some(unpack_name),
-                    Some(unpack_def),
-                    Vec::new(),
-                    Vec::new(),
-                    rw_block,
-                    false,
-                ),
-            ));
-        }
-        // Parenthesized pointy parameter list: -> ($a, $b) { ... }
-        if r.starts_with('(') {
-            let (r, _) = parse_char(r, '(')?;
-            let (r, _) = ws(r)?;
-            let (r, sub_params) = super::parse_param_list_pub(r)?;
-            let (r, _) = ws(r)?;
-            let (r, _) = parse_char(r, ')')?;
-            let (r, _) = skip_pointy_return_type(r)?;
-            if sub_params.is_empty() {
-                return Ok((r, (None, None, Vec::new(), Vec::new(), rw_block, false)));
-            }
-            let unpack_name = "__for_unpack".to_string();
-            let unpack_def = ParamDef {
-                name: unpack_name.clone(),
-                default: None,
-                multi_invocant: true,
-                required: false,
-                named: false,
-                slurpy: false,
-                sigilless: false,
-                type_constraint: None,
-                literal_value: None,
-                sub_signature: Some(sub_params),
-                where_constraint: None,
-                traits: Vec::new(),
-                double_slurpy: false,
-                onearg: false,
-                optional_marker: false,
-                outer_sub_signature: None,
-                code_signature: None,
-                is_invocant: false,
-                shape_constraints: None,
-            };
-            return Ok((
-                r,
-                (
-                    Some(unpack_name),
-                    Some(unpack_def),
-                    Vec::new(),
-                    Vec::new(),
-                    rw_block,
-                    false,
-                ),
-            ));
-        }
-        // Positional destructuring pointy param: -> [$a, $b] { ... }
-        if let Some(mut r) = r.strip_prefix('[') {
-            let (r2, _) = ws(r)?;
-            r = r2;
-            let mut sub_params = Vec::new();
-            if !r.starts_with(']') {
-                loop {
-                    let (r2, param_def) = parse_for_pointy_param(r)?;
-                    sub_params.push(param_def);
-                    let (r2, _) = ws(r2)?;
-                    if let Some(r3) = r2.strip_prefix(',') {
-                        let (r3, _) = ws(r3)?;
-                        r = r3;
-                        continue;
-                    }
-                    r = r2;
-                    break;
-                }
-            }
-            let (r, _) = parse_char(r, ']')?;
-            let (r, _) = skip_pointy_return_type(r)?;
-            let unpack_name = "__for_unpack".to_string();
-            let unpack_def = ParamDef {
-                name: unpack_name.clone(),
-                default: None,
-                multi_invocant: true,
-                required: false,
-                named: false,
-                slurpy: false,
-                sigilless: false,
-                type_constraint: None,
-                literal_value: None,
-                sub_signature: Some(sub_params),
-                where_constraint: None,
-                traits: Vec::new(),
-                double_slurpy: false,
-                onearg: false,
-                optional_marker: false,
-                outer_sub_signature: None,
-                code_signature: None,
-                is_invocant: false,
-                shape_constraints: None,
-            };
-            return Ok((
-                r,
-                (
-                    Some(unpack_name),
-                    Some(unpack_def),
-                    Vec::new(),
-                    Vec::new(),
-                    rw_block,
-                    false,
-                ),
-            ));
-        }
-        let (r, mut first_def) = parse_for_pointy_param(r)?;
-        let first = first_def.name.clone();
-        let (r, _) = ws(r)?;
-        let (r, _) = if r.starts_with('(') {
-            let (r, _) = parse_char(r, '(')?;
-            let (r, _) = ws(r)?;
-            let (r, sub_params) = super::parse_param_list_pub(r)?;
-            let (r, _) = ws(r)?;
-            let (r, _) = parse_char(r, ')')?;
-            first_def.sub_signature = Some(sub_params);
-            (r, ())
-        } else if r.starts_with('[') {
-            let (r, _) = parse_char(r, '[')?;
-            let (r, _) = ws(r)?;
-            let (r, sub_params) = super::parse_param_list_pub(r)?;
-            let (r, _) = ws(r)?;
-            let (r, _) = parse_char(r, ']')?;
-            first_def.sub_signature = Some(sub_params);
-            (r, ())
-        } else {
-            (r, ())
-        };
-        let (r, _) = ws(r)?;
-        if r.starts_with(',') {
-            let first_param = if first_def.sigilless {
-                format!("\\{}", first)
-            } else {
-                first
-            };
-            let mut params = vec![first_param];
-            let mut any_rw = rw_block || first_def.traits.iter().any(|t| t == "rw");
-            // Keep the full ParamDef per multi-param so the compiler can emit an
-            // arity check (required params) and default-value binds (`-> $a, $b = 7`).
-            let mut params_def = vec![first_def];
-            let mut r = r;
-            loop {
-                let (r2, _) = parse_char(r, ',')?;
-                let (r2, _) = ws(r2)?;
-                let (r2, next) = parse_for_pointy_param(r2)?;
-                if next.traits.iter().any(|t| t == "rw") {
-                    any_rw = true;
-                }
-                // Prefix sigilless params with \\ so the compiler can
-                // emit MarkSigillessReadonly for them.
-                if next.sigilless {
-                    params.push(format!("\\{}", next.name));
-                } else {
-                    params.push(next.name.clone());
-                }
-                params_def.push(next);
-                let (r2, _) = ws(r2)?;
-                if !r2.starts_with(',') {
-                    r = r2;
-                    break;
-                }
-                r = r2;
-            }
-            let (r, _) = skip_pointy_return_type(r)?;
-            Ok((r, (None, None, params, params_def, any_rw, false)))
-        } else {
-            let (r, _) = skip_pointy_return_type(r)?;
-            Ok((
-                r,
-                (
-                    Some(first),
-                    Some(first_def),
-                    Vec::new(),
-                    Vec::new(),
-                    rw_block,
-                    false,
-                ),
-            ))
-        }
-    } else {
-        Ok((input, (None, None, Vec::new(), Vec::new(), false, false)))
-    }
-}
-
-fn parse_for_pointy_param(input: &str) -> PResult<'_, ParamDef> {
-    // Sigilless parameter: \name
-    if let Some(r) = input.strip_prefix('\\') {
-        let (rest, name) = ident(r)?;
-        return Ok((
-            rest,
-            ParamDef {
-                name,
-                default: None,
-                multi_invocant: true,
-                required: false,
-                named: false,
-                slurpy: false,
-                double_slurpy: false,
-                onearg: false,
-                sigilless: true,
-                type_constraint: None,
-                literal_value: None,
-                sub_signature: None,
-                where_constraint: None,
-                traits: Vec::new(),
-                optional_marker: false,
-                outer_sub_signature: None,
-                code_signature: None,
-                is_invocant: false,
-                shape_constraints: None,
-            },
-        ));
-    }
-
-    let rest = input;
-    let mut type_constraint = None;
-    let rest = if let Ok((r, tc)) = ident(rest) {
-        // Preserve type smileys :D, :U, :_ as part of the type constraint.
-        let (r, tc) = if r.starts_with(":D") || r.starts_with(":U") || r.starts_with(":_") {
-            let smiley = &r[..2];
-            (&r[2..], format!("{}{}", tc, smiley))
-        } else {
-            (r, tc)
-        };
-        let (r2, _) = ws(r)?;
-        if r2.starts_with('$') || r2.starts_with('@') || r2.starts_with('%') || r2.starts_with('&')
-        {
-            type_constraint = Some(tc);
-            r2
-        } else {
-            rest
-        }
-    } else {
-        rest
-    };
-
-    let for_original_sigil = rest.as_bytes().first().copied().unwrap_or(b'$');
-    let (r, name) = var_name(rest)?;
-
-    // Shape constraint for array parameters
-    let mut shape_constraints = None;
-    let r = if for_original_sigil == b'@' && r.starts_with('[') {
-        let (r2, dims) = parse_array_shape_suffix(r)?;
-        shape_constraints = Some(dims);
-        r2
-    } else {
-        r
-    };
-
-    // `$x?` marks an optional param, `$x!` an explicitly-required one. Track the
-    // optional marker so the compiler does not count `$x?` toward the required
-    // arity of a multi-param loop.
-    let mut optional_marker = false;
-    let mut rest = if let Some(after) = r.strip_prefix('?') {
-        optional_marker = true;
-        after
-    } else if let Some(after) = r.strip_prefix('!') {
-        after
-    } else {
-        r
-    };
-
-    let mut traits = Vec::new();
-    loop {
-        let (r, _) = ws(rest)?;
-        let Some(after_is) = keyword("is", r) else {
-            rest = r;
-            break;
-        };
-        let (after_is, _) = ws1(after_is)?;
-        let (after_is, trait_name) = ident(after_is)?;
-        sub::validate_param_trait_pub(&trait_name, &traits, after_is)?;
-        traits.push(trait_name);
-        rest = after_is;
-    }
-
-    let (r, _) = ws(rest)?;
-    let mut default = None;
-    if let Some(after_eq) = r.strip_prefix('=')
-        && !after_eq.starts_with('>')
-    {
-        let (after_eq, _) = ws(after_eq)?;
-        let (after_default, default_expr) = expression(after_eq)?;
-        default = Some(default_expr);
-        rest = after_default;
-    } else {
-        rest = r;
-    }
-
-    let param_name = match for_original_sigil {
-        b'@' => format!("@{}", name),
-        b'%' => format!("%{}", name),
-        b'&' => format!("&{}", name),
-        _ => name,
-    };
-
-    Ok((
-        rest,
-        ParamDef {
-            name: param_name,
-            default,
-            multi_invocant: true,
-            required: false,
-            named: false,
-            slurpy: false,
-            double_slurpy: false,
-            onearg: false,
-            sigilless: false,
-            type_constraint,
-            literal_value: None,
-            sub_signature: None,
-            where_constraint: None,
-            traits,
-            optional_marker,
-            outer_sub_signature: None,
-            code_signature: None,
-            is_invocant: false,
-            shape_constraints,
-        },
-    ))
-}
-
-/// Reject `foreach` with X::Obsolete error (Perl 5 remnant).
-pub(super) fn foreach_stmt(input: &str) -> PResult<'_, Stmt> {
-    let rest =
-        keyword("foreach", input).ok_or_else(|| PError::expected("foreach (obsolete) check"))?;
-    // If followed by ' or - and then a letter, it's actually an identifier
-    // like `foreach'rest` or `foreach-bar`, not the obsolete keyword.
-    if let Some(rest2) = rest.strip_prefix('\'').or_else(|| rest.strip_prefix('-'))
-        && rest2.starts_with(|c: char| c.is_alphabetic() || c == '_')
-    {
-        return Err(PError::expected("foreach (obsolete) check"));
-    }
-    // `foreach(...)` or `foreach;` with no intervening whitespace could be a
-    // call of a user-defined `sub foreach {}` (see #2440). The obsolete loop
-    // form always has whitespace before its term (`foreach (1..10) { }`,
-    // `foreach @x { }`), so only bail when `(`/`;` follow immediately.
-    if rest.starts_with('(') || rest.starts_with(';') || rest.trim_start().is_empty() {
-        return Err(PError::expected("foreach (obsolete) check"));
-    }
-    Err(make_obsolete_error(
-        "'foreach'",
-        Some("'for'"),
-        "X::Obsolete: Unsupported use of 'foreach'. In Raku please use: 'for'.",
-    ))
-}
-
-/// Given input starting at `(`, return true if its matching `)` group contains
-/// a top-level `;` (the C-style `for (init; test; incr)` obsolete form).
-fn paren_has_toplevel_semicolon(input: &str) -> bool {
-    let mut depth = 0i32;
-    let mut chars = input.chars().peekable();
-    while let Some(c) = chars.next() {
-        match c {
-            '(' | '[' | '{' => depth += 1,
-            ')' | ']' | '}' => {
-                depth -= 1;
-                if depth == 0 {
-                    return false;
-                }
-            }
-            ';' if depth == 1 => return true,
-            // Skip a backslash-escaped char so an escaped delimiter is ignored.
-            '\\' => {
-                chars.next();
-            }
-            _ => {}
-        }
-    }
-    false
-}
-
 /// Build a fatal, structured `X::Obsolete` parse error carrying `old` /
 /// `replacement` attributes so that `throws-like` can match them.
 pub(in crate::parser) fn make_obsolete_error(
@@ -1093,87 +83,6 @@ pub(in crate::parser) fn make_obsolete_error(
     attrs.insert("message".to_string(), Value::str(message.to_string()));
     let exception = Value::make_instance(crate::symbol::Symbol::intern("X::Obsolete"), attrs);
     PError::fatal_with_exception(message.to_string(), Box::new(exception))
-}
-
-pub(super) fn for_stmt(input: &str) -> PResult<'_, Stmt> {
-    for_stmt_with_mode(input, crate::ast::ForMode::Normal)
-}
-
-/// Parse `race for ...` statement prefix.
-pub(super) fn race_for_stmt(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("race", input).ok_or_else(|| PError::expected("race for statement"))?;
-    let (rest, _) = ws1(rest)?;
-    for_stmt_with_mode(rest, crate::ast::ForMode::Race)
-}
-
-/// Parse `hyper for ...` statement prefix.
-pub(super) fn hyper_for_stmt(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("hyper", input).ok_or_else(|| PError::expected("hyper for statement"))?;
-    let (rest, _) = ws1(rest)?;
-    for_stmt_with_mode(rest, crate::ast::ForMode::Hyper)
-}
-
-/// Parse `for ...` with Lazy mode (input starts at `for`, not `lazy`).
-pub(super) fn lazy_for_body(input: &str) -> PResult<'_, Stmt> {
-    for_stmt_with_mode(input, crate::ast::ForMode::Lazy)
-}
-
-/// Scan for `<->` (rw pointy block) in the input, returning its byte offset.
-/// This must handle nesting to avoid matching inside strings, brackets, etc.
-fn find_rw_pointy_block(input: &str) -> Option<usize> {
-    let bytes = input.as_bytes();
-    let mut depth_paren = 0i32;
-    let mut depth_bracket = 0i32;
-    let depth_angle = 0i32;
-    let mut in_single_quote = false;
-    let mut in_double_quote = false;
-    let mut i = 0;
-    while i < bytes.len() {
-        let b = bytes[i];
-        if in_single_quote {
-            if b == b'\'' {
-                in_single_quote = false;
-            } else if b == b'\\' {
-                i += 1; // skip escaped char
-            }
-            i += 1;
-            continue;
-        }
-        if in_double_quote {
-            if b == b'"' {
-                in_double_quote = false;
-            } else if b == b'\\' {
-                i += 1;
-            }
-            i += 1;
-            continue;
-        }
-        match b {
-            b'\'' => in_single_quote = true,
-            b'"' => in_double_quote = true,
-            b'(' => depth_paren += 1,
-            b')' => depth_paren -= 1,
-            b'[' => depth_bracket += 1,
-            b']' => depth_bracket -= 1,
-            b'{' if depth_paren == 0 && depth_bracket == 0 && depth_angle == 0 => {
-                // Start of block — stop scanning
-                return None;
-            }
-            // Check for `<->`
-            b'<' if depth_paren == 0
-                && depth_bracket == 0
-                && depth_angle == 0
-                && i + 2 < bytes.len()
-                && bytes[i + 1] == b'-'
-                && bytes[i + 2] == b'>' =>
-            {
-                return Some(i);
-            }
-            _ => {}
-        }
-        i += 1;
-    }
-    None
 }
 
 /// Detect the Perl 5 `for my $x (LIST) { }` foreach pattern: a `my`/`our`/
@@ -1210,1160 +119,6 @@ fn p5_foreach_error() -> PError {
     let ex =
         crate::value::Value::make_instance(crate::symbol::Symbol::intern("X::Syntax::P5"), attrs);
     PError::fatal_with_exception(msg, Box::new(ex))
-}
-
-fn for_stmt_with_mode(input: &str, mode: crate::ast::ForMode) -> PResult<'_, Stmt> {
-    let rest = keyword("for", input).ok_or_else(|| PError::expected("for statement"))?;
-    let (rest, _) = ws1(rest)?;
-    // Perl 5 `for my $x (LIST) { }` foreach syntax.
-    if looks_like_p5_foreach(rest) {
-        return Err(p5_foreach_error());
-    }
-    // C-style `for (init; test; incr) { }` is obsolete — Raku uses `loop`.
-    // Detected by a top-level `;` inside the parenthesized iterable.
-    if rest.starts_with('(') && paren_has_toplevel_semicolon(rest) {
-        return Err(make_obsolete_error(
-            "C-style \"for\"",
-            Some("\"loop\""),
-            "X::Obsolete: Unsupported use of C-style \"for\". In Raku please use: \"loop\".",
-        ));
-    }
-    // Try to detect `<->` (rw pointy block) before the expression parser
-    // consumes the `<` as a comparison operator.
-    let (rest, iterable, rw_detected) = if let Some(rw_pos) = find_rw_pointy_block(rest) {
-        let expr_part = &rest[..rw_pos];
-        let (leftover, iterable) = parse_comma_or_expr(expr_part)?;
-        let (leftover, _) = ws(leftover)?;
-        if leftover.is_empty() {
-            (&rest[rw_pos..], iterable, true)
-        } else {
-            // Expression didn't consume everything before `<->`, fall back
-            let (r, iterable) = parse_comma_or_expr(rest)?;
-            (r, iterable, false)
-        }
-    } else {
-        let (r, iterable) = parse_comma_or_expr(rest)?;
-        (r, iterable, false)
-    };
-    let (rest, _) = ws(rest)?;
-    let (rest, (param, param_def, params, params_def, rw_block, explicit_zero_params)) =
-        parse_for_params(rest)?;
-    let rw_block = rw_block || rw_detected;
-    let (rest, _) = ws(rest)?;
-    // A `for` loop always requires a brace block. Having parsed the iterable and
-    // any pointy parameters, if no `{` follows then the block is missing
-    // (`for 1, 2` → X::Syntax::Missing, what => 'block'). Fail fatally so the
-    // statement dispatcher does not fall back to reparsing `for` as a term.
-    if !rest.starts_with('{') {
-        let mut attrs = std::collections::HashMap::new();
-        attrs.insert("what".to_string(), Value::str("block".to_string()));
-        attrs.insert(
-            "message".to_string(),
-            Value::str("Missing block".to_string()),
-        );
-        let ex = Value::make_instance(Symbol::intern("X::Syntax::Missing"), attrs);
-        return Err(PError::fatal_with_exception(
-            "X::Syntax::Missing: Missing block".to_string(),
-            Box::new(ex),
-        ));
-    }
-    // Collect &-sigil parameter names so they can be registered as user subs
-    // inside the block scope — this prevents `m()` from being misread as `m//`
-    // (a regex match) when `m` is bound via `-> &m { m() }`.
-    // Also collect sigilless `\name` params (stored with a leading `\`) so they
-    // can be registered as term symbols, shadowing any outer `&name` sub — keeps
-    // `my &mapper = {...}; for ... -> \mapper { mapper.WHAT }` reading `mapper` as
-    // the loop variable, not as a call to the outer `&mapper`.
-    let mut code_param_names: Vec<String> = Vec::new();
-    let mut sigilless_param_names: Vec<String> = Vec::new();
-    for p in param.iter().chain(params.iter()) {
-        if let Some(bare) = p.strip_prefix('&') {
-            code_param_names.push(bare.to_string());
-        } else if let Some(bare) = p.strip_prefix('\\') {
-            sigilless_param_names.push(bare.to_string());
-        }
-    }
-    // The single-param form stores a sigilless `\name` without the leading `\`
-    // (multi-param form keeps it), so consult the ParamDef sigilless flags too.
-    for pd in param_def.iter().chain(params_def.iter()) {
-        if pd.sigilless && !sigilless_param_names.contains(&pd.name) {
-            sigilless_param_names.push(pd.name.clone());
-        }
-    }
-    let (rest, body) = if !code_param_names.is_empty() || !sigilless_param_names.is_empty() {
-        let (r, _) = parse_char(rest, '{')?;
-        super::simple::push_scope();
-        for name in &code_param_names {
-            super::simple::register_user_sub(name);
-        }
-        for name in &sigilless_param_names {
-            super::simple::register_user_term_symbol(name);
-        }
-        let result = block_inner(r);
-        super::simple::pop_scope();
-        result?
-    } else {
-        block(rest)?
-    };
-    // When no explicit params, collect placeholder variables from the body
-    let (param, params) = if param.is_none() && params.is_empty() {
-        let placeholders = collect_placeholders_shallow(&body);
-        if placeholders.is_empty() {
-            (param, params)
-        } else {
-            // Use the first placeholder as the loop param, rest as extra params
-            let first = placeholders[0].clone();
-            let rest_params = placeholders[1..].to_vec();
-            (Some(first), rest_params)
-        }
-    } else {
-        (param, params)
-    };
-    Ok((
-        rest,
-        Stmt::For {
-            iterable,
-            param,
-            param_def: Box::new(param_def),
-            params,
-            params_def,
-            body,
-            label: None,
-            mode,
-            rw_block,
-            explicit_zero_params,
-        },
-    ))
-}
-
-pub(super) fn parse_pointy_param(input: &str) -> PResult<'_, ParamDef> {
-    // Sigilless parameter: \name
-    if let Some(r) = input.strip_prefix('\\') {
-        let (rest, name) = ident(r)?;
-        return Ok((
-            rest,
-            ParamDef {
-                name,
-                default: None,
-                multi_invocant: true,
-                required: false,
-                named: false,
-                slurpy: false,
-                double_slurpy: false,
-                onearg: false,
-                sigilless: true,
-                type_constraint: None,
-                literal_value: None,
-                sub_signature: None,
-                outer_sub_signature: None,
-                code_signature: None,
-                where_constraint: None,
-                traits: Vec::new(),
-                optional_marker: false,
-                is_invocant: false,
-                shape_constraints: None,
-            },
-        ));
-    }
-    // Optional type constraint before the variable
-    // Use parse_type_constraint_expr to handle coercion types (e.g., Numeric(Cool)),
-    // qualified names (Int::Odd), definedness markers (:D/:U), etc.
-    let mut rest = input;
-    let mut type_constraint = None;
-    rest = if let Some((r, tc)) = super::sub_param::parse_type_constraint_expr(rest) {
-        let (r2, _) = ws(r)?;
-        if r2.starts_with('$')
-            || r2.starts_with('@')
-            || r2.starts_with('%')
-            || r2.starts_with('&')
-            || r2.starts_with('*')
-            || (r2.starts_with(':')
-                && r2.len() > 1
-                && matches!(r2.as_bytes()[1], b'$' | b'@' | b'%' | b'&'))
-        {
-            type_constraint = Some(tc);
-            r2
-        } else if r2.starts_with('{') || r2.starts_with(',') || r2.starts_with("-->") {
-            // Type-only parameter in pointy block (e.g., `-> True { }`, `-> Int { }`)
-            let tc_for_bool = if tc == "True" || tc == "False" {
-                super::super::add_parse_warning(format!(
-                    "Potential difficulties:\n    Literal values in signatures are smartmatched against and smartmatch with `{}` will always {}. Use the `where` clause instead.",
-                    tc,
-                    if tc == "True" { "succeed" } else { "fail" }
-                ));
-                "Bool".to_string()
-            } else {
-                tc.clone()
-            };
-            return Ok((
-                r2,
-                ParamDef {
-                    name: "__type_only__".to_string(),
-                    default: None,
-                    multi_invocant: true,
-                    required: false,
-                    named: false,
-                    slurpy: false,
-                    double_slurpy: false,
-                    onearg: false,
-                    sigilless: false,
-                    type_constraint: Some(tc_for_bool),
-                    literal_value: None,
-                    sub_signature: None,
-                    outer_sub_signature: None,
-                    code_signature: None,
-                    where_constraint: None,
-                    traits: Vec::new(),
-                    optional_marker: false,
-                    is_invocant: false,
-                    shape_constraints: None,
-                },
-            ));
-        } else {
-            rest
-        }
-    } else {
-        rest
-    };
-
-    // Slurpy marker for pointy params: *@a, *%h, *$x, *&cb, and double-slurpy variants.
-    let mut slurpy = false;
-    let mut double_slurpy = false;
-    if rest.starts_with("**")
-        && rest.len() > 2
-        && matches!(rest.as_bytes()[2], b'@' | b'%' | b'$' | b'&')
-    {
-        slurpy = true;
-        double_slurpy = true;
-        rest = &rest[2..];
-    } else if rest.starts_with('*')
-        && rest.len() > 1
-        && matches!(rest.as_bytes()[1], b'@' | b'%' | b'$' | b'&')
-    {
-        slurpy = true;
-        rest = &rest[1..];
-    }
-
-    // Capture-all pointy parameter: -> |, -> |$c, -> |c
-    if let Some(stripped) = rest.strip_prefix('|') {
-        let (r, _) = ws(stripped)?;
-        // Optional capture variable name with sigil
-        if r.starts_with('$') || r.starts_with('@') || r.starts_with('%') || r.starts_with('&') {
-            let (r, name) = var_name(r)?;
-            return Ok((
-                r,
-                ParamDef {
-                    name,
-                    default: None,
-                    multi_invocant: true,
-                    required: false,
-                    named: false,
-                    slurpy: true,
-                    double_slurpy: false,
-                    onearg: false,
-                    sigilless: false,
-                    type_constraint,
-                    literal_value: None,
-                    sub_signature: None,
-                    outer_sub_signature: None,
-                    code_signature: None,
-                    where_constraint: None,
-                    traits: Vec::new(),
-                    optional_marker: false,
-                    is_invocant: false,
-                    shape_constraints: None,
-                },
-            ));
-        }
-
-        // Sigilless capture variable name
-        if let Ok((r, name)) = ident(r)
-            && !matches!(name.as_str(), "where" | "is")
-        {
-            return Ok((
-                r,
-                ParamDef {
-                    name,
-                    default: None,
-                    multi_invocant: true,
-                    required: false,
-                    named: false,
-                    slurpy: true,
-                    double_slurpy: false,
-                    onearg: false,
-                    sigilless: true,
-                    type_constraint,
-                    literal_value: None,
-                    sub_signature: None,
-                    outer_sub_signature: None,
-                    code_signature: None,
-                    where_constraint: None,
-                    traits: Vec::new(),
-                    optional_marker: false,
-                    is_invocant: false,
-                    shape_constraints: None,
-                },
-            ));
-        }
-
-        // Bare capture marker. An anonymous capture absorbs both positional and
-        // named arguments, exactly like a named capture (|c), so it must be
-        // sigilless for binding/dispatch to treat it as a full slurpy.
-        return Ok((
-            r,
-            ParamDef {
-                name: "_capture".to_string(),
-                default: None,
-                multi_invocant: true,
-                required: false,
-                named: false,
-                slurpy: true,
-                double_slurpy: false,
-                onearg: false,
-                sigilless: true,
-                type_constraint,
-                literal_value: None,
-                sub_signature: None,
-                outer_sub_signature: None,
-                code_signature: None,
-                where_constraint: None,
-                traits: Vec::new(),
-                optional_marker: false,
-                is_invocant: false,
-                shape_constraints: None,
-            },
-        ));
-    }
-
-    // Anonymous callable pointy parameter with code signature: -> &:(Str) { ... }
-    if rest.starts_with("&:(") {
-        let r = &rest[1..]; // skip '&'
-        let (r, _) = parse_char(r, ':')?;
-        let (r, _) = parse_char(r, '(')?;
-        let (r, _) = ws(r)?;
-        let (r, (sig_params, sig_ret)) = super::sub::parse_param_list_with_return(r)?;
-        let (r, _) = ws(r)?;
-        let (r, _) = parse_char(r, ')')?;
-
-        let mut optional_marker = false;
-        let mut required_marker = false;
-        let mut rest = if let Some(after) = r.strip_prefix('?') {
-            optional_marker = true;
-            after
-        } else if let Some(after) = r.strip_prefix('!') {
-            required_marker = true;
-            after
-        } else {
-            r
-        };
-
-        let mut traits = Vec::new();
-        loop {
-            let (r, _) = ws(rest)?;
-            let Some(after_is) = keyword("is", r) else {
-                rest = r;
-                break;
-            };
-            let (after_is, _) = ws1(after_is)?;
-            let (after_is, trait_name) = ident(after_is)?;
-            sub::validate_param_trait_pub(&trait_name, &traits, after_is)?;
-            traits.push(trait_name);
-            rest = after_is;
-        }
-        return Ok((
-            rest,
-            ParamDef {
-                name: "&".to_string(),
-                default: None,
-                multi_invocant: true,
-                required: required_marker,
-                named: false,
-                slurpy,
-                double_slurpy,
-                onearg: false,
-                sigilless: false,
-                type_constraint,
-                literal_value: None,
-                sub_signature: None,
-                outer_sub_signature: None,
-                code_signature: Some((sig_params, sig_ret)),
-                where_constraint: None,
-                traits,
-                optional_marker,
-                is_invocant: false,
-                shape_constraints: None,
-            },
-        ));
-    }
-
-    // Literal value parameter in pointy signatures, e.g. `-> 42 { ... }`
-    if rest.starts_with(|c: char| c.is_ascii_digit()) {
-        let mut end = 0usize;
-        for (idx, ch) in rest.char_indices() {
-            if ch.is_ascii_digit() || ch == '_' {
-                end = idx + ch.len_utf8();
-            } else {
-                break;
-            }
-        }
-        let token = &rest[..end];
-        let parsed = token.replace('_', "").parse::<i64>().unwrap_or(0);
-        let rest = &rest[end..];
-        return Ok((
-            rest,
-            ParamDef {
-                name: "__literal__".to_string(),
-                default: None,
-                multi_invocant: true,
-                required: false,
-                named: false,
-                slurpy,
-                double_slurpy,
-                onearg: false,
-                sigilless: false,
-                type_constraint,
-                literal_value: Some(crate::value::Value::Int(parsed)),
-                sub_signature: None,
-                outer_sub_signature: None,
-                code_signature: None,
-                where_constraint: None,
-                traits: Vec::new(),
-                optional_marker: false,
-                is_invocant: false,
-                shape_constraints: None,
-            },
-        ));
-    }
-
-    // Named parameter prefix: :$x, :@l, :%h
-    let mut named = false;
-    if rest.starts_with(':')
-        && rest.len() > 1
-        && matches!(rest.as_bytes()[1], b'$' | b'@' | b'%' | b'&')
-    {
-        named = true;
-        rest = &rest[1..];
-    }
-
-    let original_sigil = rest.as_bytes().first().copied().unwrap_or(b'$');
-    let (rest, name) = var_name(rest)?;
-
-    // Shape constraint for array parameters: @a[3], @a[*]
-    let mut shape_constraints = None;
-    let rest = if original_sigil == b'@' && rest.starts_with('[') {
-        let (r, dims) = parse_array_shape_suffix(rest)?;
-        shape_constraints = Some(dims);
-        r
-    } else {
-        rest
-    };
-
-    // Optional marker on pointy params: $x? / $x!
-    let mut optional_marker = false;
-    let mut required_marker = false;
-    let mut rest = if let Some(after) = rest.strip_prefix('?') {
-        optional_marker = true;
-        after
-    } else if let Some(after) = rest.strip_prefix('!') {
-        required_marker = true;
-        after
-    } else {
-        rest
-    };
-
-    // Optional parameter traits: `is rw`, `is copy`, ...
-    let mut traits = Vec::new();
-    loop {
-        let (r, _) = ws(rest)?;
-        let Some(after_is) = keyword("is", r) else {
-            rest = r;
-            break;
-        };
-        let (after_is, _) = ws1(after_is)?;
-        let (after_is, trait_name) = ident(after_is)?;
-        sub::validate_param_trait_pub(&trait_name, &traits, after_is)?;
-        traits.push(trait_name);
-        rest = after_is;
-    }
-
-    // Optional default value: `$x = expr`
-    let mut default = None;
-    let (r, _) = ws(rest)?;
-    if let Some(after_eq) = r.strip_prefix('=')
-        && !after_eq.starts_with('>')
-    {
-        let (after_eq, _) = ws(after_eq)?;
-        let (after_default, default_expr) = expression(after_eq)?;
-        default = Some(default_expr);
-        rest = after_default;
-    } else {
-        rest = r;
-    }
-
-    // Optional unpacking sub-signature: `-> Pair $p (:$key, :$value) { ... }`
-    // Parse and preserve the sub-signature for runtime binding.
-    let mut sub_signature = None;
-    let (r, _) = ws(rest)?;
-    if r.starts_with('(') {
-        let (r, _) = parse_char(r, '(')?;
-        let (r, _) = ws(r)?;
-        let (r, sub_params) = super::sub::parse_param_list(r)?;
-        let (r, _) = ws(r)?;
-        let (r, _) = parse_char(r, ')')?;
-        sub_signature = Some(sub_params);
-        rest = r;
-    } else {
-        rest = r;
-    }
-
-    // Optional where-constraint on pointy parameters: `-> $x where Int|Bool { ... }`
-    let (r, _) = ws(rest)?;
-    let (rest, where_constraint) = if let Some(r) = keyword("where", r) {
-        let (r, _) = ws1(r)?;
-        let (r, constraint) = if r.starts_with('{') {
-            let (r, body) = crate::parser::primary::parse_block_body(r)?;
-            (
-                r,
-                Expr::AnonSub {
-                    body,
-                    is_rw: false,
-                    is_block: true,
-                },
-            )
-        } else {
-            expression(r)?
-        };
-        (r, Some(Box::new(constraint)))
-    } else {
-        (r, None)
-    };
-
-    // Prefix name with sigil for proper runtime binding
-    let param_name = match original_sigil {
-        b'@' => format!("@{}", name),
-        b'%' => format!("%{}", name),
-        b'&' => format!("&{}", name),
-        _ => name,
-    };
-    Ok((
-        rest,
-        ParamDef {
-            name: param_name,
-            default,
-            multi_invocant: true,
-            required: required_marker,
-            named,
-            slurpy,
-            double_slurpy,
-            onearg: false,
-            sigilless: false,
-            type_constraint,
-            literal_value: None,
-            sub_signature,
-            outer_sub_signature: None,
-            code_signature: None,
-            where_constraint,
-            traits,
-            optional_marker,
-            is_invocant: false,
-            shape_constraints,
-        },
-    ))
-}
-
-/// Parse `while` loop.
-pub(super) fn while_stmt(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("while", input).ok_or_else(|| PError::expected("while statement"))?;
-    let (rest, _) = ws1(rest)?;
-    let (rest, cond) = condition_expr(rest)?;
-    let (rest, _) = ws(rest)?;
-    let (rest, param_binding) = if rest.starts_with("->") {
-        let (rest, (param, _param_def, params, _params_def, _rw_block, _explicit_zero)) =
-            parse_for_params(rest)?;
-        if !params.is_empty() {
-            return Err(PError::expected_at("single while pointy parameter", rest));
-        }
-        (rest, param)
-    } else {
-        (rest, None::<String>)
-    };
-    let (rest, _) = ws(rest)?;
-    let (rest, body) = block(rest)?;
-    let (hoisted_decl, cond) = if param_binding.is_none() {
-        split_loop_cond_decl(cond)
-    } else {
-        (None, cond)
-    };
-    let while_stmt = Stmt::While {
-        cond: if let Some(ref param) = param_binding {
-            Expr::AssignExpr {
-                name: param.clone(),
-                expr: Box::new(cond),
-                is_bind: false,
-            }
-        } else {
-            cond
-        },
-        body,
-        label: None,
-    };
-    if let Some(decl) = hoisted_decl {
-        return Ok((rest, Stmt::Block(vec![decl, while_stmt])));
-    }
-    if let Some(param) = param_binding {
-        Ok((
-            rest,
-            Stmt::Block(vec![
-                Stmt::VarDecl {
-                    name: param,
-                    expr: Expr::Literal(crate::value::Value::Nil),
-                    type_constraint: None,
-                    is_state: false,
-                    is_our: false,
-                    is_dynamic: false,
-                    is_export: false,
-                    export_tags: Vec::new(),
-                    custom_traits: Vec::new(),
-                    where_constraint: None,
-                },
-                while_stmt,
-            ]),
-        ))
-    } else {
-        Ok((rest, while_stmt))
-    }
-}
-
-/// Parse `until` loop.
-pub(super) fn until_stmt(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("until", input).ok_or_else(|| PError::expected("until statement"))?;
-    let (rest, _) = ws1(rest)?;
-    let (rest, cond) = condition_expr(rest)?;
-    let (rest, _) = ws(rest)?;
-    let (rest, param_binding) = if rest.starts_with("->") {
-        let (rest, (param, _param_def, params, _params_def, _rw_block, _explicit_zero)) =
-            parse_for_params(rest)?;
-        if !params.is_empty() {
-            return Err(PError::expected_at("single until pointy parameter", rest));
-        }
-        (rest, param)
-    } else {
-        (rest, None::<String>)
-    };
-    let (rest, _) = ws(rest)?;
-    let (rest, body) = block(rest)?;
-    let (hoisted_decl, cond) = if param_binding.is_none() {
-        split_loop_cond_decl(cond)
-    } else {
-        (None, cond)
-    };
-    let cond_expr = if let Some(ref param) = param_binding {
-        Expr::AssignExpr {
-            name: param.clone(),
-            expr: Box::new(cond),
-            is_bind: false,
-        }
-    } else {
-        cond
-    };
-    let while_stmt = Stmt::While {
-        cond: Expr::Unary {
-            op: TokenKind::Bang,
-            expr: Box::new(cond_expr),
-        },
-        body,
-        label: None,
-    };
-    if let Some(decl) = hoisted_decl {
-        return Ok((rest, Stmt::Block(vec![decl, while_stmt])));
-    }
-    if let Some(param) = param_binding {
-        Ok((
-            rest,
-            Stmt::Block(vec![
-                Stmt::VarDecl {
-                    name: param,
-                    expr: Expr::Literal(crate::value::Value::Nil),
-                    type_constraint: None,
-                    is_state: false,
-                    is_our: false,
-                    is_dynamic: false,
-                    is_export: false,
-                    export_tags: Vec::new(),
-                    custom_traits: Vec::new(),
-                    where_constraint: None,
-                },
-                while_stmt,
-            ]),
-        ))
-    } else {
-        Ok((rest, while_stmt))
-    }
-}
-
-/// Parse a single init/step item for a C-style loop header.
-/// In loop headers, commas separate statements, so `my $a = 1, my $b = 2`
-/// must parse as two separate declarations, not one list assignment.
-/// This function parses `my`/`state` declarations using `expression()` (not
-/// `parse_assign_expr_or_comma()`) for the RHS, preventing comma consumption.
-fn loop_header_item(input: &str) -> PResult<'_, Stmt> {
-    let (input, _) = ws(input)?;
-    // Try simple my/state scalar declarations with non-comma-consuming RHS.
-    // For `loop (my $a = 1, my $b = 2; ...)`, commas must separate
-    // declarations, not be part of the assignment value.
-    if keyword("my", input).is_some() || keyword("state", input).is_some() {
-        let is_state = keyword("state", input).is_some();
-        let rest = keyword("my", input)
-            .or_else(|| keyword("state", input))
-            .unwrap();
-        let (rest, _) = ws1(rest)?;
-        // Only handle simple scalar declarations (no type constraint).
-        // Typed declarations like `my int $i` fall through to statement().
-        if let Ok((rest, name)) = var_name(rest) {
-            // var_name strips the sigil; VarDecl name does not include the sigil
-            let (rest, _) = ws(rest)?;
-            if rest.starts_with('=') && !rest.starts_with("==") && !rest.starts_with("=>") {
-                let rest = &rest[1..];
-                let (rest, _) = ws(rest)?;
-                let (rest, expr) = expression(rest)?;
-                return Ok((
-                    rest,
-                    Stmt::VarDecl {
-                        name,
-                        expr,
-                        type_constraint: None,
-                        is_state,
-                        is_our: false,
-                        is_dynamic: false,
-                        is_export: false,
-                        export_tags: Vec::new(),
-                        custom_traits: Vec::new(),
-                        where_constraint: None,
-                    },
-                ));
-            }
-            return Ok((
-                rest,
-                Stmt::VarDecl {
-                    name,
-                    expr: Expr::Literal(crate::value::Value::Nil),
-                    type_constraint: None,
-                    is_state,
-                    is_our: false,
-                    is_dynamic: false,
-                    is_export: false,
-                    export_tags: Vec::new(),
-                    custom_traits: Vec::new(),
-                    where_constraint: None,
-                },
-            ));
-        }
-        // Typed or complex declaration (e.g., my int $i = 0) — fall through
-        // to statement() which handles the full my_decl grammar.
-        return statement(input);
-    }
-    // Fall back to expression statement (also non-comma-consuming)
-    let (rest, expr) = expression(input)?;
-    Ok((rest, Stmt::Expr(expr)))
-}
-
-/// Build a typed X::Syntax::Malformed error for a bad C-style loop spec.
-/// `what` is the `loop spec (...)` description matched by the roast tests.
-fn malformed_loop_spec_error(what: &str) -> PError {
-    let message = format!("X::Syntax::Malformed: Malformed {}.", what);
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("message".to_string(), Value::str(message.clone()));
-    attrs.insert("what".to_string(), Value::str(what.to_string()));
-    let exception = Value::make_instance(Symbol::intern("X::Syntax::Malformed"), attrs);
-    PError::fatal_with_exception(message, Box::new(exception))
-}
-
-/// Return the slice from just-after `loop (` up to (excluding) the matching `)`,
-/// honoring nested brackets and quotes, or None if unbalanced.
-fn loop_spec_content(after_open_paren: &str) -> Option<&str> {
-    let bytes = after_open_paren.as_bytes();
-    let mut depth = 0i32;
-    let mut quote: Option<u8> = None;
-    let mut i = 0;
-    while i < bytes.len() {
-        let c = bytes[i];
-        if let Some(q) = quote {
-            if c == b'\\' {
-                i += 2;
-                continue;
-            }
-            if c == q {
-                quote = None;
-            }
-        } else {
-            match c {
-                b'\'' | b'"' => quote = Some(c),
-                b'(' | b'[' | b'{' => depth += 1,
-                b']' | b'}' => depth -= 1,
-                b')' => {
-                    if depth == 0 {
-                        return Some(&after_open_paren[..i]);
-                    }
-                    depth -= 1;
-                }
-                _ => {}
-            }
-        }
-        i += 1;
-    }
-    None
-}
-
-/// Count the top-level `;`-separated groups in a C-style loop spec (respecting
-/// nested brackets/quotes). A valid spec has exactly 3.
-fn loop_spec_semicolon_groups(spec: &str) -> usize {
-    let bytes = spec.as_bytes();
-    let mut depth = 0i32;
-    let mut quote: Option<u8> = None;
-    let mut semis = 0usize;
-    let mut i = 0;
-    while i < bytes.len() {
-        let c = bytes[i];
-        if let Some(q) = quote {
-            if c == b'\\' {
-                i += 2;
-                continue;
-            }
-            if c == q {
-                quote = None;
-            }
-        } else {
-            match c {
-                b'\'' | b'"' => quote = Some(c),
-                b'(' | b'[' | b'{' => depth += 1,
-                b')' | b']' | b'}' => depth -= 1,
-                b';' if depth == 0 => semis += 1,
-                _ => {}
-            }
-        }
-        i += 1;
-    }
-    semis + 1
-}
-
-/// Validate a C-style loop spec: exactly 3 semicolon-separated expressions.
-/// Returns an X::Syntax::Malformed error (with the right `what`) otherwise.
-fn check_loop_spec(spec: &str) -> Option<PError> {
-    if spec.trim().is_empty() {
-        return Some(malformed_loop_spec_error(
-            "loop spec (expected 3 semicolon-separated expressions)",
-        ));
-    }
-    let groups = loop_spec_semicolon_groups(spec);
-    if groups == 3 {
-        return None;
-    }
-    let what = if groups > 3 {
-        "loop spec (expected 3 semicolon-separated expressions but got more)".to_string()
-    } else {
-        format!("loop spec (expected 3 semicolon-separated expressions but got {groups})")
-    };
-    Some(malformed_loop_spec_error(&what))
-}
-
-/// Parse C-style `loop` or infinite loop.
-pub(super) fn loop_stmt(input: &str) -> PResult<'_, Stmt> {
-    let after_kw = keyword("loop", input).ok_or_else(|| PError::expected("loop statement"))?;
-    let (rest, _) = ws(after_kw)?;
-    // Whitespace between `loop` and `(` marks a real C-style loop spec; `loop(`
-    // with no space is a keyword-as-function mistake handled elsewhere, so the
-    // spec validation below must NOT preempt it.
-    let had_space_before_paren = rest.len() < after_kw.len();
-    if rest.starts_with('(') {
-        let (rest, _) = parse_char(rest, '(')?;
-        // Reject a malformed loop spec (not exactly 3 `;`-separated expressions)
-        // with a typed X::Syntax::Malformed before the lenient parse below.
-        if had_space_before_paren
-            && let Some(spec) = loop_spec_content(rest)
-            && let Some(err) = check_loop_spec(spec)
-        {
-            return Err(err);
-        }
-        let (rest, _) = ws(rest)?;
-        // init: may be comma-separated list of statements
-        let (rest, init) = if let Some(rest_after_semi) = rest.strip_prefix(';') {
-            (rest_after_semi, None)
-        } else {
-            let (mut r, first) = loop_header_item(rest)?;
-            let mut stmts = vec![first];
-            loop {
-                let (r2, _) = ws(r)?;
-                if let Some(r2_after_comma) = r2.strip_prefix(',') {
-                    let (r2, _) = ws(r2_after_comma)?;
-                    let (r2, s) = loop_header_item(r2)?;
-                    stmts.push(s);
-                    r = r2;
-                } else {
-                    r = r2;
-                    break;
-                }
-            }
-            if stmts.len() == 1 {
-                (r, Some(Box::new(stmts.into_iter().next().unwrap())))
-            } else {
-                (r, Some(Box::new(Stmt::SyntheticBlock(stmts))))
-            }
-        };
-        let (rest, _) = ws(rest)?;
-        // When init was parsed by loop_header_item, the semicolon separating
-        // init from cond is not yet consumed. Consume it now.
-        // When init is None, the semicolon was already consumed by strip_prefix(';').
-        let rest = if init.is_some() {
-            if let Some(r) = rest.strip_prefix(';') {
-                r
-            } else {
-                rest
-            }
-        } else {
-            rest
-        };
-        let (rest, _) = ws(rest)?;
-        // cond
-        let (rest, cond) = if let Some(rest) = rest.strip_prefix(';') {
-            (rest, None)
-        } else {
-            let (r, e) = expression(rest)?;
-            let (r, _) = ws(r)?;
-            let (r, _) = parse_char(r, ';')?;
-            (r, Some(e))
-        };
-        let (rest, _) = ws(rest)?;
-        // step: may be comma-separated list of expressions
-        let (rest, step) = if rest.starts_with(')') {
-            (rest, None)
-        } else {
-            let (mut r, first) = expression(rest)?;
-            let mut exprs = vec![first];
-            loop {
-                let (r2, _) = ws(r)?;
-                if let Some(r2_after_comma) = r2.strip_prefix(',') {
-                    let (r2, _) = ws(r2_after_comma)?;
-                    let (r2, e) = expression(r2)?;
-                    exprs.push(e);
-                    r = r2;
-                } else {
-                    r = r2;
-                    break;
-                }
-            }
-            if exprs.len() == 1 {
-                (r, Some(exprs.into_iter().next().unwrap()))
-            } else {
-                // Wrap multiple step expressions in a DoBlock
-                let stmts: Vec<Stmt> = exprs.into_iter().map(Stmt::Expr).collect();
-                (
-                    r,
-                    Some(Expr::DoBlock {
-                        body: stmts,
-                        label: None,
-                    }),
-                )
-            }
-        };
-        let (rest, _) = ws(rest)?;
-        let (rest, _) = parse_char(rest, ')')?;
-        let (rest, _) = ws(rest)?;
-        let (rest, body) = block(rest)?;
-        return Ok((
-            rest,
-            Stmt::Loop {
-                init,
-                cond,
-                step,
-                body,
-                repeat: false,
-                label: None,
-            },
-        ));
-    }
-    // Infinite loop
-    let (rest, _) = ws(rest)?;
-    let (rest, body) = block(rest)?;
-    // `loop {} while COND` is a syntax error — use `repeat while` instead
-    let (check, _) = ws(rest)?;
-    if keyword("while", check).is_some() || keyword("until", check).is_some() {
-        return Err(PError::fatal(
-            "X::Syntax::Confused: Strange text after block (missing semicolon or comma?)"
-                .to_string(),
-        ));
-    }
-    Ok((
-        rest,
-        Stmt::Loop {
-            init: None,
-            cond: None,
-            step: None,
-            body,
-            repeat: false,
-            label: None,
-        },
-    ))
-}
-
-/// Parse `repeat while/until` loop.
-pub(super) fn repeat_stmt(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("repeat", input).ok_or_else(|| PError::expected("repeat statement"))?;
-    let (rest, _) = ws(rest)?;
-
-    // repeat while/until COND { BODY }
-    if let Some(r) = keyword("while", rest) {
-        let (r, _) = ws1(r)?;
-        let (r, cond) = condition_expr(r)?;
-        let (r, _) = ws(r)?;
-        let (r, (param, _param_def, params, _params_def, _rw_block, _explicit_zero)) =
-            parse_for_params(r)?;
-        let (r, _) = ws(r)?;
-        let (r, body) = block(r)?;
-        let (r, _) = ws(r)?;
-        let (r, _) = opt_char(r, ';');
-        let repeat_param = param.or_else(|| params.into_iter().next());
-        let init = repeat_param.as_ref().map(|name| {
-            Box::new(Stmt::VarDecl {
-                name: name.clone(),
-                expr: Expr::Literal(crate::value::Value::Nil),
-                type_constraint: None,
-                is_state: false,
-                is_our: false,
-                is_dynamic: false,
-                is_export: false,
-                export_tags: Vec::new(),
-                custom_traits: Vec::new(),
-                where_constraint: None,
-            })
-        });
-        let step = repeat_param.map(|name| Expr::AssignExpr {
-            name,
-            expr: Box::new(Expr::Literal(crate::value::Value::Bool(true))),
-            is_bind: false,
-        });
-        return Ok((
-            r,
-            Stmt::Loop {
-                init,
-                cond: Some(cond),
-                step,
-                body,
-                repeat: true,
-                label: None,
-            },
-        ));
-    }
-    if let Some(r) = keyword("until", rest) {
-        let (r, _) = ws1(r)?;
-        let (r, cond) = condition_expr(r)?;
-        let (r, _) = ws(r)?;
-        let (r, (param, _param_def, params, _params_def, _rw_block, _explicit_zero)) =
-            parse_for_params(r)?;
-        let (r, _) = ws(r)?;
-        let (r, body) = block(r)?;
-        let (r, _) = ws(r)?;
-        let (r, _) = opt_char(r, ';');
-        let repeat_param = param.or_else(|| params.into_iter().next());
-        let init = repeat_param.as_ref().map(|name| {
-            Box::new(Stmt::VarDecl {
-                name: name.clone(),
-                expr: Expr::Literal(crate::value::Value::Nil),
-                type_constraint: None,
-                is_state: false,
-                is_our: false,
-                is_dynamic: false,
-                is_export: false,
-                export_tags: Vec::new(),
-                custom_traits: Vec::new(),
-                where_constraint: None,
-            })
-        });
-        let step = repeat_param.map(|name| Expr::AssignExpr {
-            name,
-            expr: Box::new(Expr::Literal(crate::value::Value::Bool(true))),
-            is_bind: false,
-        });
-        return Ok((
-            r,
-            Stmt::Loop {
-                init,
-                cond: Some(Expr::Unary {
-                    op: TokenKind::Bang,
-                    expr: Box::new(cond),
-                }),
-                step,
-                body,
-                repeat: true,
-                label: None,
-            },
-        ));
-    }
-
-    // repeat { BODY } while/until COND
-    let (rest, body) = block(rest)?;
-    let (rest, _) = ws(rest)?;
-    if let Some(r) = keyword("while", rest) {
-        let (r, _) = ws1(r)?;
-        let (r, cond) = condition_expr(r)?;
-        let (r, _) = ws(r)?;
-        let (r, _) = opt_char(r, ';');
-        return Ok((
-            r,
-            Stmt::Loop {
-                init: None,
-                cond: Some(cond),
-                step: None,
-                body,
-                repeat: true,
-                label: None,
-            },
-        ));
-    }
-    if let Some(r) = keyword("until", rest) {
-        let (r, _) = ws1(r)?;
-        let (r, cond) = condition_expr(r)?;
-        let (r, _) = ws(r)?;
-        let (r, _) = opt_char(r, ';');
-        return Ok((
-            r,
-            Stmt::Loop {
-                init: None,
-                cond: Some(Expr::Unary {
-                    op: TokenKind::Bang,
-                    expr: Box::new(cond),
-                }),
-                step: None,
-                body,
-                repeat: true,
-                label: None,
-            },
-        ));
-    }
-    Err(PError::fatal(
-        "X::Syntax::Missing: \"while\" or \"until\" required after repeat".to_string(),
-    ))
-}
-
-/// Parse `given`/`when`/`default`.
-pub(super) fn given_stmt(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("given", input).ok_or_else(|| PError::expected("given statement"))?;
-    let (rest, _) = ws1(rest)?;
-    let (rest, topic) = expression(rest)?;
-    let (rest, _) = ws(rest)?;
-    // Check for pointy block: given EXPR -> $param { ... }
-    let (rest, pointy_param) = if let Some(r) = rest.strip_prefix("->") {
-        let (r, _) = ws(r)?;
-        let (r, pd) = parse_pointy_param(r)?;
-        let (r, _) = ws(r)?;
-        (r, Some(pd))
-    } else {
-        (rest, None)
-    };
-    let (rest, mut body) = block(rest)?;
-    if let Some(pd) = pointy_param {
-        body.insert(0, pointy_topic_bind(&pd));
-    }
-    Ok((rest, Stmt::Given { topic, body }))
 }
 
 /// Build the head statement that binds a pointy-block parameter to the topic.
@@ -2414,22 +169,6 @@ fn method_call(target: Expr, name: &str) -> Expr {
     }
 }
 
-pub(super) fn when_stmt(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("when", input).ok_or_else(|| PError::expected("when statement"))?;
-    let (rest, _) = ws1(rest)?;
-    let (rest, cond) = condition_expr(rest)?;
-    let (rest, _) = ws(rest)?;
-    let (rest, body) = block(rest)?;
-    Ok((rest, Stmt::When { cond, body }))
-}
-
-pub(super) fn default_stmt(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("default", input).ok_or_else(|| PError::expected("default statement"))?;
-    let (rest, _) = ws(rest)?;
-    let (rest, body) = block(rest)?;
-    Ok((rest, Stmt::Default(body)))
-}
-
 /// Build a `$_ = <expr>` assignment statement for topicalization.
 fn topicalize(expr: &Expr) -> Stmt {
     Stmt::Assign {
@@ -2439,503 +178,31 @@ fn topicalize(expr: &Expr) -> Stmt {
     }
 }
 
-/// Parse `with`/`without` statement.
-pub(super) fn with_stmt(input: &str) -> PResult<'_, Stmt> {
-    let is_without = keyword("without", input).is_some();
-    let rest = keyword("with", input)
-        .or_else(|| keyword("without", input))
-        .ok_or_else(|| PError::expected("with/without statement"))?;
-    let (rest, _) = ws1(rest)?;
-    let (rest, cond_expr) = condition_expr(rest)?;
-    let (rest, _) = ws(rest)?;
+mod conditionals;
+mod for_loops;
+mod for_params;
+mod given_when;
+mod labeled_loop;
+mod loop_repeat;
+mod pointy_param;
+mod react;
+mod while_until;
+mod with_stmt;
 
-    // Check for optional pointy block: -> $param { ... }, -> \param { ... },
-    // or -> (SIGNATURE) { ... } (sub-signature destructuring)
-    let (rest, param_name, param_def) = if rest.starts_with("->") || rest.starts_with("<->") {
-        let (r, (param, param_def, _params, _params_def, _rw_block, _explicit_zero)) =
-            parse_for_params(rest)?;
-        (r, param, param_def)
-    } else {
-        (rest, None, None)
-    };
+// Re-exports preserving each function's original visibility (all `pub(super)`).
+pub(super) use conditionals::{if_stmt, unless_stmt};
+pub(super) use for_loops::{for_stmt, foreach_stmt, hyper_for_stmt, lazy_for_body, race_for_stmt};
+pub(super) use for_params::parse_for_params;
+pub(super) use given_when::{default_stmt, given_stmt, when_stmt};
+pub(super) use labeled_loop::labeled_loop_stmt;
+pub(super) use loop_repeat::{loop_stmt, repeat_stmt};
+pub(super) use pointy_param::parse_pointy_param;
+pub(super) use react::{react_stmt, whenever_stmt};
+pub(super) use while_until::{until_stmt, while_stmt};
+pub(super) use with_stmt::with_stmt;
 
-    let (rest, body) = block(rest)?;
-
-    // Use a temp variable in the condition to evaluate cond_expr exactly once.
-    // This is important for expressions like `Failure.new` where evaluating
-    // twice would create two distinct objects, and the `.defined` call on the
-    // condition would not mark the same instance that `$_` receives.
-    static WITH_COUNTER: std::sync::atomic::AtomicUsize = std::sync::atomic::AtomicUsize::new(0);
-    let with_id = WITH_COUNTER.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-    let tmp_name = format!("__with_tmp_{}", with_id);
-    let tmp_var = Expr::Var(tmp_name.clone());
-
-    // The body uses $_ which was set in the condition block.
-    // We still prepend topicalize(&tmp_var) as a no-op marker so that
-    // $_ is visible in the body scope (but it's already set by the condition).
-    // If the topic source is a literal value, wrap $_ in a Mixin marked as
-    // read-only so that mutating ops like tr/// throw X::Assignment::RO.
-    // Encoding the read-only flag in the value itself avoids leaking marker
-    // variables across exception boundaries.
-    // When the topic is a plain variable and there is no pointy parameter, run
-    // the body under a `given` so `$_` aliases that variable (matching `given`'s
-    // semantics). This lets in-place mutations through `$_` (e.g. `.=uc`,
-    // `$_ = ...`) write back to the original container, while `$_` stays writable
-    // exactly as in `given`. The `.defined` condition is still evaluated
-    // separately via `$tmp`. Container variables (`@a` / `%h`) also alias their
-    // container topic through `given`, so `with @a { .push }` propagates the
-    // mutation back to `@a` (the `given` opcode treats `@`/`%` topics as
-    // read-only-for-reassign but writes container mutations back to the source).
-    // A container *element* subscript on a simple variable (`with %h<k>` /
-    // `with @a[i]`) is an lvalue: routing it through `given` makes the `given`
-    // element-source writeback propagate both `$_ = ...` and `.push` back to the
-    // element (the element topic is rw, unlike a whole-container topic).
-    let topic_is_element_lvalue = matches!(
-        &cond_expr,
-        Expr::Index { target, .. }
-            if matches!(
-                target.as_ref(),
-                Expr::Var(_) | Expr::ArrayVar(_) | Expr::HashVar(_)
-            )
-    );
-    let cond_is_lvalue = matches!(
-        &cond_expr,
-        Expr::Var(_) | Expr::ArrayVar(_) | Expr::HashVar(_)
-    ) || topic_is_element_lvalue;
-    // A container/scalar pointy parameter on an lvalue condition is routed
-    // through `given` so it shares `given`'s topic semantics: a plain `-> @p`
-    // aliases the source (writeback of `@p.push` / `@p[0]=v`), while `-> @p is
-    // copy` becomes a flattened fresh copy (see `pointy_topic_bind`). Exclude
-    // sub-signatures, sigilless `\x`, callables, and attributive (`$!x`/`$.x`)
-    // params — those keep the copy-style VarDecl binding below.
-    let pointy_routes_through_given = param_def
-        .as_ref()
-        .map(|pd| {
-            pd.sub_signature.is_none()
-                && !pd.sigilless
-                && !pd.name.is_empty()
-                && !pd.name.starts_with('&')
-                && !pd.name.starts_with('!')
-                && !pd.name.starts_with('.')
-        })
-        .unwrap_or(false);
-    let use_given_alias = cond_is_lvalue && (param_name.is_none() || pointy_routes_through_given);
-    // Topicalize `$_` for the body. For a literal, wrap it in a Mixin marked
-    // read-only so in-place mutation of `$_` throws X::Assignment::RO. Otherwise
-    // reuse the `$tmp` holding the (once-evaluated) condition value.
-    let topic_assign = if let Expr::Literal(lit) = &cond_expr {
-        let mut overrides = std::collections::HashMap::new();
-        overrides.insert("__mutsu_topic_ro__".to_string(), Value::Bool(true));
-        let wrapped = Value::mixin(lit.clone(), overrides);
-        Stmt::Assign {
-            name: "_".to_string(),
-            op: AssignOp::Assign,
-            expr: Expr::Literal(wrapped),
-        }
-    } else {
-        topicalize(&tmp_var)
-    };
-    let mut with_body = if use_given_alias {
-        Vec::new()
-    } else {
-        vec![topic_assign]
-    };
-    // If a named parameter was given (-> $param), also assign it. A
-    // container/scalar pointy param is handled by the `given` routing below
-    // (it prepends the alias/copy bind to the given body), so skip the VarDecl
-    // here for those.
-    if let Some(ref pname) = param_name
-        && !(use_given_alias && pointy_routes_through_given)
-    {
-        // Attributive parameter (`-> $!foo` / `-> $.foo`): bind the value to
-        // self's attribute via an attribute assignment. This must be checked
-        // first (an attributive param never has a sub-signature), and it must be
-        // an Assign (not a VarDecl) so it writes through to the attribute even
-        // when the body runs inside a nested `given` topic scope.
-        if pname.starts_with('!') || pname.starts_with('.') {
-            let attr_name = &pname[1..];
-            with_body.push(Stmt::Assign {
-                name: format!("!{}", attr_name),
-                expr: tmp_var.clone(),
-                op: crate::ast::AssignOp::Assign,
-            });
-        } else if let Some(ref pdef) = param_def {
-            if let Some(ref sub_params) = pdef.sub_signature {
-                // Declare the unpack variable holding the condition value
-                with_body.push(Stmt::VarDecl {
-                    name: pname.clone(),
-                    expr: tmp_var.clone(),
-                    type_constraint: None,
-                    is_state: false,
-                    is_our: false,
-                    is_dynamic: false,
-                    is_export: false,
-                    export_tags: Vec::new(),
-                    custom_traits: Vec::new(),
-                    where_constraint: None,
-                });
-                // Destructure each positional parameter from the unpack variable
-                let mut positional_index = 0usize;
-                for sub in sub_params {
-                    if sub.name.is_empty() {
-                        continue;
-                    }
-                    let extract_expr = if sub.named {
-                        Expr::MethodCall {
-                            target: Box::new(Expr::Var(pname.clone())),
-                            name: Symbol::intern(&sub.name),
-                            args: Vec::new(),
-                            modifier: None,
-                            quoted: false,
-                        }
-                    } else {
-                        let idx_expr = Expr::Index {
-                            target: Box::new(Expr::Var(pname.clone())),
-                            index: Box::new(Expr::Literal(Value::Int(positional_index as i64))),
-                            is_positional: true,
-                        };
-                        positional_index += 1;
-                        idx_expr
-                    };
-                    // Apply type coercion if present (e.g. Int())
-                    let coerced_expr = if let Some(ref tc) = sub.type_constraint {
-                        // Strip trailing "()" from coercion types like "Int()"
-                        let method_name = tc.strip_suffix("()").unwrap_or(tc);
-                        Expr::MethodCall {
-                            target: Box::new(extract_expr),
-                            name: Symbol::intern(method_name),
-                            args: Vec::new(),
-                            modifier: None,
-                            quoted: false,
-                        }
-                    } else {
-                        extract_expr
-                    };
-                    // `is copy` on a sub-signature parameter is handled
-                    // implicitly: the VarDecl creates a fresh writable variable,
-                    // so no explicit trait is needed.
-                    with_body.push(Stmt::VarDecl {
-                        name: sub.name.clone(),
-                        expr: coerced_expr,
-                        type_constraint: None,
-                        is_state: false,
-                        is_our: false,
-                        is_dynamic: false,
-                        is_export: false,
-                        export_tags: Vec::new(),
-                        custom_traits: Vec::new(),
-                        where_constraint: None,
-                    });
-                }
-            } else {
-                // Simple parameter with possible traits from parse_for_params
-                with_body.push(Stmt::VarDecl {
-                    name: pname.clone(),
-                    expr: tmp_var.clone(),
-                    type_constraint: None,
-                    is_state: false,
-                    is_our: false,
-                    is_dynamic: false,
-                    is_export: false,
-                    export_tags: Vec::new(),
-                    custom_traits: Vec::new(),
-                    where_constraint: None,
-                });
-            }
-        } else {
-            with_body.push(Stmt::VarDecl {
-                name: pname.clone(),
-                expr: tmp_var.clone(),
-                type_constraint: None,
-                is_state: false,
-                is_our: false,
-                is_dynamic: false,
-                is_export: false,
-                export_tags: Vec::new(),
-                custom_traits: Vec::new(),
-                where_constraint: None,
-            });
-        }
-    }
-    if use_given_alias {
-        let mut given_body = body;
-        // For a container/scalar pointy param, prepend the alias/copy bind
-        // (`@p := $_` for aliasing, `@p = $_.list` for `is copy`) so the `given`
-        // mechanism handles topic-source writeback (alias) or a flattened fresh
-        // copy (matching `given @a -> @p`).
-        if pointy_routes_through_given && let Some(ref pd) = param_def {
-            given_body.insert(0, pointy_topic_bind(pd));
-        }
-        with_body = vec![Stmt::Given {
-            topic: cond_expr.clone(),
-            body: given_body,
-        }];
-    } else {
-        with_body.extend(body);
-    }
-
-    // Build the condition as (my $tmp = cond_expr).defined() (or
-    // !(my $tmp = cond_expr).defined() for without). The DoStmt(VarDecl)
-    // evaluates cond_expr once, declares $tmp in the current scope, and
-    // returns the value. Then .defined() checks it. The body uses $tmp
-    // for topicalization instead of re-evaluating cond_expr.
-    let var_decl_expr = Expr::DoStmt(Box::new(Stmt::VarDecl {
-        name: tmp_name.clone(),
-        expr: cond_expr.clone(),
-        type_constraint: None,
-        is_state: false,
-        is_our: false,
-        is_dynamic: false,
-        is_export: false,
-        export_tags: Vec::new(),
-        custom_traits: Vec::new(),
-        where_constraint: None,
-    }));
-    let defined_check = Expr::MethodCall {
-        target: Box::new(var_decl_expr),
-        name: Symbol::intern("defined"),
-        args: Vec::new(),
-        modifier: None,
-        quoted: false,
-    };
-    let cond = if is_without {
-        Expr::Unary {
-            op: TokenKind::Bang,
-            expr: Box::new(defined_check),
-        }
-    } else {
-        defined_check
-    };
-    // Parse orwith / else chains
-    let rest_before_ws = rest;
-    let (rest, _) = ws(rest)?;
-
-    // `without` cannot have `else`, `orwith`, or `elsif` chains (Raku spec)
-    if is_without {
-        for kw in &["else", "orwith", "elsif"] {
-            if keyword(kw, rest).is_some() {
-                return Err(PError::fatal(format!(
-                    "X::Syntax::WithoutElse: '{}' is not allowed on 'without'",
-                    kw
-                )));
-            }
-        }
-    }
-
-    let (rest, else_branch) = if keyword("orwith", rest).is_some() {
-        let r = keyword("orwith", rest).unwrap();
-        let (r, _) = ws1(r)?;
-        let (r, orwith_cond_expr) = condition_expr(r)?;
-        let (r, _) = ws(r)?;
-
-        // Check for optional pointy block on orwith: orwith EXPR -> $param { ... }
-        let (r, orwith_param_name) = if let Some(r2) = r.strip_prefix("->") {
-            let (r2, _) = ws(r2)?;
-            if let Some(r_after_sigil) = r2.strip_prefix('$') {
-                let end = r_after_sigil
-                    .find(|c: char| !c.is_alphanumeric() && c != '_' && c != '-')
-                    .unwrap_or(r_after_sigil.len());
-                let name = &r_after_sigil[..end];
-                let r2 = &r_after_sigil[end..];
-                let (r2, _) = ws(r2)?;
-                (r2, Some(name.to_string()))
-            } else if let Some(r_after_backslash) = r2.strip_prefix('\\') {
-                let end = r_after_backslash
-                    .find(|c: char| !c.is_alphanumeric() && c != '_' && c != '-')
-                    .unwrap_or(r_after_backslash.len());
-                let name = &r_after_backslash[..end];
-                let r2 = &r_after_backslash[end..];
-                let (r2, _) = ws(r2)?;
-                (r2, Some(name.to_string()))
-            } else {
-                (r, None)
-            }
-        } else {
-            (r, None)
-        };
-
-        let (r, orwith_body) = block(r)?;
-
-        // Prepend $_ = <orwith_cond_expr> to the body
-        let mut orwith_with_body = vec![topicalize(&orwith_cond_expr)];
-        if let Some(ref pname) = orwith_param_name {
-            orwith_with_body.push(Stmt::VarDecl {
-                name: pname.clone(),
-                expr: orwith_cond_expr.clone(),
-                type_constraint: None,
-                is_state: false,
-                is_our: false,
-                is_dynamic: false,
-                is_export: false,
-                export_tags: Vec::new(),
-                custom_traits: Vec::new(),
-                where_constraint: None,
-            });
-        }
-        orwith_with_body.extend(orwith_body);
-
-        let orwith_cond_expr_clone = orwith_cond_expr.clone();
-        let orwith_cond = Expr::MethodCall {
-            target: Box::new(orwith_cond_expr),
-            name: Symbol::intern("defined"),
-            args: Vec::new(),
-            modifier: None,
-            quoted: false,
-        };
-        let r_before_else_ws = r;
-        let (r, _) = ws(r)?;
-        let (r, orwith_else) = if keyword("else", r).is_some() {
-            let r2 = keyword("else", r).unwrap();
-            let (r2, _) = ws(r2)?;
-            // Check for optional pointy block on else: else -> $param { ... }
-            let (r2, else_param) = if let Some(r3) = r2.strip_prefix("->") {
-                let (r3, _) = ws(r3)?;
-                if let Some(r_after_sigil) = r3.strip_prefix('$') {
-                    let end = r_after_sigil
-                        .find(|c: char| !c.is_alphanumeric() && c != '_' && c != '-')
-                        .unwrap_or(r_after_sigil.len());
-                    let name = &r_after_sigil[..end];
-                    let r3 = &r_after_sigil[end..];
-                    let (r3, _) = ws(r3)?;
-                    (r3, Some(name.to_string()))
-                } else {
-                    (r2, None)
-                }
-            } else {
-                (r2, None)
-            };
-            let (r2, else_body) = block(r2)?;
-            // Topicalize $_ in else branch to the orwith condition
-            let mut else_with_topic = vec![topicalize(&orwith_cond_expr_clone)];
-            if let Some(ref pname) = else_param {
-                else_with_topic.push(Stmt::VarDecl {
-                    name: pname.clone(),
-                    expr: orwith_cond_expr_clone.clone(),
-                    type_constraint: None,
-                    is_state: false,
-                    is_our: false,
-                    is_dynamic: false,
-                    is_export: false,
-                    export_tags: Vec::new(),
-                    custom_traits: Vec::new(),
-                    where_constraint: None,
-                });
-            }
-            else_with_topic.extend(else_body);
-            (r2, else_with_topic)
-        } else {
-            (r_before_else_ws, Vec::new())
-        };
-        (
-            r,
-            vec![Stmt::If {
-                cond: orwith_cond,
-                then_branch: orwith_with_body,
-                else_branch: orwith_else,
-                binding_var: None,
-            }],
-        )
-    } else if keyword("elsif", rest).is_some() {
-        // Handle elsif/else chains (reuse if-chain infrastructure)
-        let (r, (elsif_clauses, else_clause)) = parse_elsif_chain(rest)?;
-        let elsif_stmt = lower_if_chain(elsif_clauses, else_clause);
-        (r, vec![elsif_stmt])
-    } else if keyword("else", rest).is_some() {
-        let r = keyword("else", rest).unwrap();
-        let (r, _) = ws(r)?;
-        // Check for optional pointy block on else: else -> $param { ... }
-        let (r, else_param) = if let Some(r2) = r.strip_prefix("->") {
-            let (r2, _) = ws(r2)?;
-            if let Some(r_after_sigil) = r2.strip_prefix('$') {
-                let end = r_after_sigil
-                    .find(|c: char| !c.is_alphanumeric() && c != '_' && c != '-')
-                    .unwrap_or(r_after_sigil.len());
-                let name = &r_after_sigil[..end];
-                let r2 = &r_after_sigil[end..];
-                let (r2, _) = ws(r2)?;
-                (r2, Some(name.to_string()))
-            } else {
-                (r, None)
-            }
-        } else {
-            (r, None)
-        };
-        let (r, else_body) = block(r)?;
-        // Topicalize $_ in else branch to the with/without condition
-        let mut else_with_topic = vec![topicalize(&tmp_var)];
-        if let Some(ref pname) = else_param {
-            else_with_topic.push(Stmt::VarDecl {
-                name: pname.clone(),
-                expr: tmp_var.clone(),
-                type_constraint: None,
-                is_state: false,
-                is_our: false,
-                is_dynamic: false,
-                is_export: false,
-                export_tags: Vec::new(),
-                custom_traits: Vec::new(),
-                where_constraint: None,
-            });
-        }
-        else_with_topic.extend(else_body);
-        (r, else_with_topic)
-    } else {
-        // No orwith/else found — don't consume whitespace/newlines past the
-        // closing brace, so the caller sees the statement boundary correctly
-        // (e.g. `do with X { ... }\nsay ...` should not absorb `say`).
-        (rest_before_ws, Vec::new())
-    };
-
-    Ok((
-        rest,
-        Stmt::If {
-            cond,
-            then_branch: with_body,
-            else_branch,
-            binding_var: None,
-        },
-    ))
-}
-
-/// Parse `react` block.
-/// Supports both `react { ... }` and `react whenever ... { ... }` (shorthand).
-pub(super) fn react_stmt(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("react", input).ok_or_else(|| PError::expected("react block"))?;
-    let (rest, _) = ws(rest)?;
-    // Try `react whenever ...` shorthand first
-    if let Ok((rest2, whenever)) = whenever_stmt(rest) {
-        return Ok((
-            rest2,
-            Stmt::React {
-                body: vec![whenever],
-            },
-        ));
-    }
-    let (rest, body) = block(rest)?;
-    Ok((rest, Stmt::React { body }))
-}
-
-/// Parse `whenever` block.
-pub(super) fn whenever_stmt(input: &str) -> PResult<'_, Stmt> {
-    let rest = keyword("whenever", input).ok_or_else(|| PError::expected("whenever block"))?;
-    let (rest, _) = ws1(rest)?;
-    let (rest, supply) = expression(rest)?;
-    let (rest, _) = ws(rest)?;
-    let (rest, param) = if let Some(stripped) = rest.strip_prefix("->") {
-        let (r, _) = ws(stripped)?;
-        let (r, name) = var_name(r)?;
-        (r, Some(name))
-    } else {
-        (rest, None)
-    };
-    let (rest, _) = ws(rest)?;
-    let (rest, body) = block(rest)?;
-    Ok((
-        rest,
-        Stmt::Whenever {
-            supply,
-            param,
-            body,
-        },
-    ))
-}
+// Cross-submodule helpers: originally private `fn`s in the if/unless block that
+// are also called from `with_stmt` (for `orwith` elsif-chain lowering). Bumped
+// to `pub(crate)` in `conditionals` and re-exported here so `with_stmt` reaches
+// them via `use super::*`.
+pub(in crate::parser) use conditionals::{lower_if_chain, parse_elsif_chain};
